Poprawa ocen Włączyny -> włączony ilość -> liczba

+++ app/src/main/java/com/example/remotelightoperator/plantdescription/PlantFullDescriptionActivity.java
@@ -46,7 +46,7 @@ public class PlantFullDescriptionActivity extends Activity {
         description.setText(plantTemplate.getDescription());
         irradiationTime.setText(String.valueOf(plantTemplate.getIrradiationTime()));
         rate.setText(String.valueOf(plantTemplate.getRate()));
-        rateCount.setText(String.valueOf(plantTemplate.getRateCount()));
+        rateCount.setText(String.valueOf((int)plantTemplate.getRateCount()));
 
         RatingBar ratingBar = (RatingBar) findViewById(R.id.ratingBar);
         TextView ratingText = (TextView) findViewById(R.id.rateText);
